remove events foldrer
