+++ src/features/index.ts
@@ -1,7 +1,13 @@
 import { Plugin } from '@/types/Plugin'
 
+import memberUpdateLogger from './memberUpdateLogger'
 import nameChecker from './nameChecker'
 import preventEmojiSpam from './preventEmojiSpam'
 import stickyMessage from './stickyMessage'
 
-export default [nameChecker, preventEmojiSpam, stickyMessage] as Plugin[]
+export default [
+  memberUpdateLogger,
+  nameChecker,
+  preventEmojiSpam,
+  stickyMessage,
+] as Plugin[]
